requirements.txt
aiogram
requests
python-dotenv
psycopg2-binary
